payment: invoke control plugin state machine in Janitor GET flow
This fixes #1017.
